+++ spec/integrations/rails/rails5_pristine/railtie_setup_spec.rb
@@ -2,6 +2,9 @@
 
 # Karafka should work with Rails 5 using the default setup
 
+# Load all the Railtie stuff like when `rails server`
+ENV['KARAFKA_CLI'] = 'true'
+
 Bundler.require(:default)
 
 require 'tempfile'