+++ spec/integrations/rails/rails7_pristine/just-a-dependency/rails_setup_spec.rb
@@ -1,17 +1,13 @@
 # frozen_string_literal: true
 
-# Karafka should work fine when someone has root level components named based on the features
+# Karafka should work fine with Rails 7 even when it is just a transitive dependency and is not
+# in active use. In case like this KARAFKA_BOOT_FILE needs to be set to "false"
 #
-# @see https://github.com/karafka/karafka/issues/1144
+# @see https://github.com/karafka/karafka/issues/813
 
-class Builder
-end
+# Load all the Railtie stuff like when `rails server`
 
-class Topic
-end
-
-class Contract
-end
+ENV['KARAFKA_CLI'] = 'true'
 
 Bundler.require(:default)